requirements.txt
asgiref==3.9.1
bleach==6.2.0
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.6
django-crispy-forms==2.4
django-summernote==0.8.20.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
webencodings==0.5.1
whitenoise==6.11.0
